requirements.txt
sc2reader==1.8.0
black==24.2.0
sqlalchemy==2.0.28
pytest==8.1.1
Flask==3.0.3
pygetwindow==0.0.9
flask-cors==4.0.0
psutil==5.9.8
pyqt==5.15.10